:sparkles: [Feat] 프로젝트 초기세팅
