+++ src/main/java/com/simter/.github/pull_request_template.md
@@ -1,0 +1,11 @@
+## #️⃣ 관련 이슈
+- ex) #이슈번호
+
+## ✍️ 작업 내용 요약
+> 작업 내용 요약 
+
+## 💡 작업내용
+- 작업 배경을 알려주세요.
+
+
+## 🏞 기타
