--- src/main/java/simter/simter/SimterApplication.java
@@ -1,13 +1,0 @@
-package simter.simter;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class SimterApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(SimterApplication.class, args);
-	}
-
-}
